Fix bug with MacOS default Terminal
The default terminal application that comes with the MacOS sends a tap on
the mousepad as rapid-fire left-button down/left-button up.  When the
get_input() method reads in the key buffer, it ends up with both codes
in the buffer at once.  Due to the check for a buffer with a mouse
code being 6 bytes (and this double-code containing 12), the mouse
code was rejected.  As a result, you couldn't click on anything in the
Mac terminal.

This hacky fix just allows the double-code.  The left-button up is
consumed in the process, but since the framework only pays attention
to the left-button down, this loss isn't important.